Enable source maps in webpack
Without it, browsers may misinterpret source map data and throw an
ERR_UNKNOWN_URL_SCHEME in their dev tools console. This can be
resolved be setting the devtools mode in the webpack config,
which will then have webpack properly generate maps to the contents
of the combined file back to its unbuilt state for devtools to
utilize.