Introduce user authentication via JWT
It is now possible to login, logout, and register new users. More
work may need to be done (e.g. this CL does not handle the case
of expired refresh tokens) but the site is now user-accessible!

I also implemented a React hook `useAuthAxios()` for handling access
token refresh when the user agent needs to access a protected API
endpoint, all with the aid of the `axios` npm package.