fix: Only serve from the host machine rather than any host
Running tests would have WebSocket handshakes fail because the origin
was not specified. A workaround was to just allow any host (thereby
also allowing a host with a "None" value in its Origin header). This
is of course a security concern.

The solution is for each function test case (at least those that
use WebsocketCommunicator) to pass an origin header. With this, we
can specify in ALLOWED_HOSTS to just allow hosts associated to the
host machine.