feat(chat): Implement the `group:fetch` event
This event exists so that so that a client can request all member
and message data for a particular chat group that they are a member
of.

The reason why this isn't done after a WebSocket connection is
established is so that we won't grab every message and member list
from every chat group the user is apart of, which is an approach
that won't scale well.