feat(chat): Implement `group:message` event
Clients can now send this event when they wish to send a new message
to a chat group they are a member of. This message will be saved
to the database and sent to the channel layer group associated
to the chat group, where the channels involved will then send this
new message to their client.