feat(users): Add friends list view endpoint
Debated whether this should just be another WebSocket event in the
`chat` app or an API endpoint in `users`. I decided for now to have
it just be an endpoint as having a client fetch their friends list
does not require real-time communication and it makes sense for it
to belong here.